requirements.txt
virl2-client
urllib3
PyYAML




